Updates
I fixed the mobile view and added a tablet view of 800px, changed up the color of the background of the About Us border. I also adjusted some height for the About Us border to make it all fit. Still trying to figure out on how to put my own images in the react-images library that I am using.